Fix issue in delete method.
Previously, when encountering a ProtectedError in the delete method, only one protected object was being returned with an error message.
However, according to Django's delete behavior, all protected objects should be returned with a message.
This commit addresses this issue by modifying the delete method to return all protected objects along with the appropriate message.